Fix: Fixed delete method.

+++ src/main/java/com/example/demo/service/TechnicalInterviewService.java
@@ -1,19 +1,29 @@
 package com.example.demo.service;
 
 import com.example.demo.entity.TechnicalInterviewEntity;
+import com.example.demo.exception.QuestionNotFoundException;
 import com.example.demo.repository.TechnicalInterviewRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TechnicalInterviewService {
 
-    @Autowired
-    TechnicalInterviewRepository technicalInterviewRepository;
+    final TechnicalInterviewRepository technicalInterviewRepository;
 
+    @Value("${question.error.message}")
+    private String notFoundErrorMessage;
+
+    @Value("${question.info.message}")
+    private String notFoundInfoMessage;
+
+    public TechnicalInterviewService(TechnicalInterviewRepository technicalInterviewRepository) {
+        this.technicalInterviewRepository = technicalInterviewRepository;
+    }
 
     @Transactional
     public List<TechnicalInterviewEntity> getAllTechnicalInterviewTasks() {
@@ -22,12 +32,15 @@ public class TechnicalInterviewService {
 
     @Transactional
     public List<TechnicalInterviewEntity> getRandomQuestion() {
+        List<TechnicalInterviewEntity> list = technicalInterviewRepository.findQuestion();
+        if (list.isEmpty() || list.size() == 0) {
+            throw new QuestionNotFoundException(notFoundInfoMessage);
+        }
         return technicalInterviewRepository.findQuestion();
     }
 
     @Transactional
     public TechnicalInterviewEntity saveTechnicalInterviewTask(TechnicalInterviewEntity technicalInterviewEntity) {
-
         return technicalInterviewRepository.save(technicalInterviewEntity);
     }
 
@@ -49,6 +62,10 @@ public class TechnicalInterviewService {
 
     @Transactional
     public void deleteTechnicalInterviewTask(final Long id) {
+        Optional<TechnicalInterviewEntity> technicalInterviewEntity = technicalInterviewRepository.findById(id);
+        if (technicalInterviewEntity.isEmpty()) {
+            throw new QuestionNotFoundException(notFoundErrorMessage);
+        }
         technicalInterviewRepository.deleteById(id);
     }
